+++ CHANGELOG.md
@@ -10,3 +10,9 @@
 
 * Created core http client for RESTful communication
 * Extended and created Contentful and Management clients for Contentful APIs
+
+## 0.2.0
+
+* Created Example project with UI that runs `Flutter` to test package.
+* Fixed bug where API responses were 404.
+* Fixed bug where Authorization was missing from request headers.